Using the custom precision in the extensions

+++ src/synthesizer/extensions/doppler_particle_spectra.cpp
@@ -34,7 +34,7 @@
  * Note: binary search returns the index of the upper bin of those that straddle
  * the given lambda.
  */
-int get_upper_lam_bin(double lambda, double *grid_wavelengths, int nlam) {
+int get_upper_lam_bin(FLOAT lambda, FLOAT *grid_wavelengths, int nlam) {
   return binary_search(0, nlam - 1, grid_wavelengths, lambda);
 }
 
@@ -51,13 +51,13 @@ int get_upper_lam_bin(double lambda, double *grid_wavelengths, int nlam) {
  */
 static void shifted_spectra_loop_cic_serial(GridProps *grid_props,
                                             Particles *parts,
-                                            double *part_spectra,
-                                            const double c) {
+                                            FLOAT *part_spectra,
+                                            const FLOAT c) {
 
   /* Unpack the grid properties. */
   const int ndim = grid_props->ndim;
   size_t nlam = static_cast<size_t>(grid_props->nlam);
-  double *wavelength = grid_props->get_lam();
+  FLOAT *wavelength = grid_props->get_lam();
   const int ncells = 1 << ndim;
 
   /* Get and cast the number of particles. */
@@ -85,7 +85,7 @@ static void shifted_spectra_loop_cic_serial(GridProps *grid_props,
   }
 
   /* Allocate arrays for shifted wavelengths and bin mappings once */
-  std::vector<double> shifted_wavelengths(nlam);
+  std::vector<FLOAT> shifted_wavelengths(nlam);
   std::vector<int> mapped_indices(nlam);
 
   /* Loop over particles. */
@@ -97,12 +97,12 @@ static void shifted_spectra_loop_cic_serial(GridProps *grid_props,
     }
 
     /* Compute shift factor for this particle. */
-    const double vel = parts->get_vel_at(p);
-    const double shift_factor = 1.0 + vel / c;
+    const FLOAT vel = parts->get_vel_at(p);
+    const FLOAT shift_factor = 1.0 + vel / c;
 
     /* Shift wavelengths & map to bins once per particle. */
     for (size_t il = 0; il < nlam; ++il) {
-      const double lam_s = wavelength[il] * shift_factor;
+      const FLOAT lam_s = wavelength[il] * shift_factor;
       shifted_wavelengths[il] = lam_s;
       mapped_indices[il] = get_upper_lam_bin(lam_s, wavelength, nlam);
     }
@@ -111,14 +111,14 @@ static void shifted_spectra_loop_cic_serial(GridProps *grid_props,
     const int base_lin = parts->grid_indices[p];
 
     /* Cache particle weight once */
-    const double w_p = parts->get_weight_at(p);
+    const FLOAT w_p = parts->get_weight_at(p);
 
     /* Loop over all 2^ndim sub-cells. */
     for (int ic = 0; ic < ncells; ++ic) {
       const auto &sc = subcells[ic];
 
       /* Compute CIC fraction for this corner */
-      double frac = 1.0;
+      FLOAT frac = 1.0;
       for (int d = 0; d < ndim; ++d) {
         frac *= sc.offs[d] ? parts->grid_fracs[p * ndim + d]
                            : (1.0 - parts->grid_fracs[p * ndim + d]);
@@ -127,7 +127,7 @@ static void shifted_spectra_loop_cic_serial(GridProps *grid_props,
         continue;
 
       /* Combined weight */
-      const double weight = frac * w_p;
+      const FLOAT weight = frac * w_p;
 
       /* Flattened grid index */
       const int grid_i = base_lin + sc.linoff;
@@ -143,12 +143,12 @@ static void shifted_spectra_loop_cic_serial(GridProps *grid_props,
         }
 
         /* Interpolation fraction between bins */
-        const double lam_s = shifted_wavelengths[il];
-        const double frac_s = (lam_s - wavelength[ils - 1]) /
+        const FLOAT lam_s = shifted_wavelengths[il];
+        const FLOAT frac_s = (lam_s - wavelength[ils - 1]) /
                               (wavelength[ils] - wavelength[ils - 1]);
 
         /* Base spectra value */
-        const double gs = grid_props->get_spectra_at(grid_i, il) * weight;
+        const FLOAT gs = grid_props->get_spectra_at(grid_i, il) * weight;
 
         /* Distribute into particle & global arrays */
         const size_t base_idx = p * nlam;
@@ -180,13 +180,13 @@ static void shifted_spectra_loop_cic_serial(GridProps *grid_props,
  */
 #ifdef WITH_OPENMP
 static void shifted_spectra_loop_cic_omp(GridProps *grid_props,
-                                         Particles *parts, double *part_spectra,
-                                         int nthreads, const double c) {
+                                         Particles *parts, FLOAT *part_spectra,
+                                         int nthreads, const FLOAT c) {
 
   /* Unpack the grid properties. */
   const int ndim = grid_props->ndim;
   size_t nlam = static_cast<size_t>(grid_props->nlam);
-  double *wavelength = grid_props->get_lam();
+  FLOAT *wavelength = grid_props->get_lam();
   const int ncells = 1 << ndim;
 
   /* Get and cast the number of particles. */
@@ -216,7 +216,7 @@ static void shifted_spectra_loop_cic_omp(GridProps *grid_props,
 #pragma omp parallel num_threads(nthreads)
   {
     /* Allocate per-thread shift buffers once */
-    std::vector<double> shifted_wavelengths(nlam);
+    std::vector<FLOAT> shifted_wavelengths(nlam);
     std::vector<int> mapped_indices(nlam);
 
     /* Split the work evenly across threads (no single particle is more
@@ -232,10 +232,10 @@ static void shifted_spectra_loop_cic_omp(GridProps *grid_props,
         (tid == nthreads - 1) ? parts->npart : start_idx + nparts_per_thread;
 
     /* Get this threads part of the output array. */
-    double *__restrict local_part_spectra = part_spectra + start_idx * nlam;
+    FLOAT *__restrict local_part_spectra = part_spectra + start_idx * nlam;
 
     /* Get an array that we'll put each particle's spectra into. */
-    std::vector<double> this_part_spectra(nlam, 0.0);
+    std::vector<FLOAT> this_part_spectra(nlam, 0.0);
 
     /* Loop over particles in this thread's range. */
     for (size_t p = start_idx; p < end_idx; p++) {
@@ -246,26 +246,26 @@ static void shifted_spectra_loop_cic_omp(GridProps *grid_props,
       }
 
       /* Compute the Doppler shift factor. */
-      const double vel = parts->get_vel_at(p);
-      const double shift_factor = 1.0 + vel / c;
+      const FLOAT vel = parts->get_vel_at(p);
+      const FLOAT shift_factor = 1.0 + vel / c;
 
       /* Shift wavelengths & map to bins once per particle */
       for (size_t il = 0; il < nlam; ++il) {
-        const double lam_s = wavelength[il] * shift_factor;
+        const FLOAT lam_s = wavelength[il] * shift_factor;
         shifted_wavelengths[il] = lam_s;
         mapped_indices[il] = get_upper_lam_bin(lam_s, wavelength, nlam);
       }
 
       /* Compute base linear index and cached weight */
       const int base_lin = parts->grid_indices[p];
-      const double w_p = parts->get_weight_at(p);
+      const FLOAT w_p = parts->get_weight_at(p);
 
       /* Loop over all 2^ndim sub-cells */
       for (int ic = 0; ic < ncells; ++ic) {
         const auto &sc = subcells[ic];
 
         /* Compute CIC fraction for this corner */
-        double frac = 1.0;
+        FLOAT frac = 1.0;
         for (int d = 0; d < ndim; ++d) {
           frac *= sc.offs[d] ? parts->grid_fracs[p * ndim + d]
                              : (1.0 - parts->grid_fracs[p * ndim + d]);
@@ -274,7 +274,7 @@ static void shifted_spectra_loop_cic_omp(GridProps *grid_props,
           continue;
 
         /* Combined weight */
-        const double weight = frac * w_p;
+        const FLOAT weight = frac * w_p;
         const int grid_i = base_lin + sc.linoff;
 
         /* Loop over wavelengths (we can't prepare the unmasked wavelengths like
@@ -288,12 +288,12 @@ static void shifted_spectra_loop_cic_omp(GridProps *grid_props,
           }
 
           /* Interpolation fraction */
-          const double lam_s = shifted_wavelengths[il];
-          const double frac_s = (lam_s - wavelength[ils - 1]) /
+          const FLOAT lam_s = shifted_wavelengths[il];
+          const FLOAT frac_s = (lam_s - wavelength[ils - 1]) /
                                 (wavelength[ils] - wavelength[ils - 1]);
 
           /* Base spectra contribution */
-          const double gs = grid_props->get_spectra_at(grid_i, il) * weight;
+          const FLOAT gs = grid_props->get_spectra_at(grid_i, il) * weight;
 
           /* Deposit into the thread's part spectra */
           this_part_spectra[ils - 1] =
@@ -324,8 +324,8 @@ static void shifted_spectra_loop_cic_omp(GridProps *grid_props,
  * @param nthreads: The number of threads to use.
  */
 void shifted_spectra_loop_cic(GridProps *grid_props, Particles *parts,
-                              double *part_spectra, const int nthreads,
-                              const double c) {
+                              FLOAT *part_spectra, const int nthreads,
+                              const FLOAT c) {
 
   /* First get the grid indices and fractions for all particles. */
   get_particle_indices_and_fracs(grid_props, parts, nthreads);
@@ -368,18 +368,18 @@ void shifted_spectra_loop_cic(GridProps *grid_props, Particles *parts,
  */
 static void shifted_spectra_loop_ngp_serial(GridProps *grid_props,
                                             Particles *parts,
-                                            double *part_spectra,
-                                            const double c) {
+                                            FLOAT *part_spectra,
+                                            const FLOAT c) {
 
   /* Unpack the grid properties. */
   size_t nlam = static_cast<size_t>(grid_props->nlam);
-  double *wavelength = grid_props->get_lam();
+  FLOAT *wavelength = grid_props->get_lam();
 
   /* Get and cast the number of particles. */
   size_t npart = static_cast<size_t>(parts->npart);
 
   /* Allocate the shifted wavelengths array and the mapped indices array. */
-  std::vector<double> shifted_wavelengths(nlam);
+  std::vector<FLOAT> shifted_wavelengths(nlam);
   std::vector<int> mapped_indices(nlam);
 
   /* Loop over particles. */
@@ -391,18 +391,18 @@ static void shifted_spectra_loop_ngp_serial(GridProps *grid_props,
     }
 
     /* Get the particle velocity and red/blue shift factor. */
-    double vel = parts->get_vel_at(p);
-    double shift_factor = 1.0 + vel / c;
+    FLOAT vel = parts->get_vel_at(p);
+    FLOAT shift_factor = 1.0 + vel / c;
 
     /* Shift wavelengths & map to bins once per particle. */
     for (size_t il = 0; il < nlam; ++il) {
-      const double lam_s = wavelength[il] * shift_factor;
+      const FLOAT lam_s = wavelength[il] * shift_factor;
       shifted_wavelengths[il] = lam_s;
       mapped_indices[il] = get_upper_lam_bin(lam_s, wavelength, nlam);
     }
 
     /* Define the weight. */
-    double weight = parts->get_weight_at(p);
+    FLOAT weight = parts->get_weight_at(p);
 
     /* Get the weight's index. */
     const int grid_ind = parts->grid_indices[p];
@@ -414,7 +414,7 @@ static void shifted_spectra_loop_ngp_serial(GridProps *grid_props,
 
       /* Get the shifted wavelength and index. */
       int ilam_shifted = mapped_indices[ilam];
-      double shifted_lambda = shifted_wavelengths[ilam];
+      FLOAT shifted_lambda = shifted_wavelengths[ilam];
 
       /* Skip if this wavelength is masked. */
       if (grid_props->lam_is_masked(ilam_shifted)) {
@@ -423,7 +423,7 @@ static void shifted_spectra_loop_ngp_serial(GridProps *grid_props,
 
       /* Compute the fraction of the shifted wavelength between the two
        * closest wavelength elements. */
-      double frac_shifted = 0.0;
+      FLOAT frac_shifted = 0.0;
       if (ilam_shifted > 0 && static_cast<size_t>(ilam_shifted) <= nlam - 1) {
         frac_shifted =
             (shifted_lambda - wavelength[ilam_shifted - 1]) /
@@ -434,7 +434,7 @@ static void shifted_spectra_loop_ngp_serial(GridProps *grid_props,
       }
 
       /* Get the grid spectra value for this wavelength. */
-      double grid_spectra_value =
+      FLOAT grid_spectra_value =
           grid_props->get_spectra_at(grid_ind, ilam) * weight;
 
       /* Add the contribution to the corresponding wavelength element. */
@@ -457,12 +457,12 @@ static void shifted_spectra_loop_ngp_serial(GridProps *grid_props,
  */
 #ifdef WITH_OPENMP
 static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
-                                         Particles *parts, double *part_spectra,
-                                         int nthreads, const double c) {
+                                         Particles *parts, FLOAT *part_spectra,
+                                         int nthreads, const FLOAT c) {
 
   /* Unpack the grid properties. */
   size_t nlam = static_cast<size_t>(grid_props->nlam);
-  double *wavelength = grid_props->get_lam();
+  FLOAT *wavelength = grid_props->get_lam();
 
   /* Get and cast the number of particles. */
   size_t npart = static_cast<size_t>(parts->npart);
@@ -471,7 +471,7 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
   {
 
     /* Allocate the shifted wavelengths array and the mapped indices array. */
-    std::vector<double> shifted_wavelengths(nlam);
+    std::vector<FLOAT> shifted_wavelengths(nlam);
     std::vector<int> mapped_indices(nlam);
 
     /* Split the work evenly across threads (no single particle is more
@@ -487,10 +487,10 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
         (tid == nthreads - 1) ? parts->npart : start_idx + nparts_per_thread;
 
     /* Get this threads part of the output array. */
-    double *__restrict local_part_spectra = part_spectra + start_idx * nlam;
+    FLOAT *__restrict local_part_spectra = part_spectra + start_idx * nlam;
 
     /* Get an array that we'll put each particle's spectra into. */
-    std::vector<double> this_part_spectra(nlam, 0.0);
+    std::vector<FLOAT> this_part_spectra(nlam, 0.0);
 
     /* Loop over particles in this thread's range. */
     for (size_t p = start_idx; p < end_idx; p++) {
@@ -501,18 +501,18 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
       }
 
       /* Get the particle velocity and red/blue shift factor. */
-      double vel = parts->get_vel_at(p);
-      double shift_factor = 1.0 + vel / c;
+      FLOAT vel = parts->get_vel_at(p);
+      FLOAT shift_factor = 1.0 + vel / c;
 
       /* Shift wavelengths & map to bins once per particle. */
       for (size_t il = 0; il < nlam; ++il) {
-        const double lam_s = wavelength[il] * shift_factor;
+        const FLOAT lam_s = wavelength[il] * shift_factor;
         shifted_wavelengths[il] = lam_s;
         mapped_indices[il] = get_upper_lam_bin(lam_s, wavelength, nlam);
       }
 
       /* Define the weighted contribution from this cell. */
-      const double weight = parts->get_weight_at(p);
+      const FLOAT weight = parts->get_weight_at(p);
 
       /* Get the index of the grid cell. */
       const int grid_ind = parts->grid_indices[p];
@@ -524,7 +524,7 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
 
         /* Get the shifted wavelength and index. */
         int ilam_shifted = mapped_indices[ilam];
-        double shifted_lambda = shifted_wavelengths[ilam];
+        FLOAT shifted_lambda = shifted_wavelengths[ilam];
 
         /* Skip if this wavelength is masked. */
         if (grid_props->lam_is_masked(ilam_shifted)) {
@@ -533,7 +533,7 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
 
         /* Compute the fraction of the shifted wavelength between the two
          * closest wavelength elements. */
-        double frac_shifted = 0.0;
+        FLOAT frac_shifted = 0.0;
         if (ilam_shifted > 0 && static_cast<size_t>(ilam_shifted) <= nlam - 1) {
           frac_shifted =
               (shifted_lambda - wavelength[ilam_shifted - 1]) /
@@ -544,7 +544,7 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
         }
 
         /* Get the grid spectra value for this wavelength. */
-        double grid_spectra_value =
+        FLOAT grid_spectra_value =
             grid_props->get_spectra_at(grid_ind, ilam) * weight;
 
         /* Deposit into the thread's part spectra */
@@ -557,7 +557,7 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
 
       /* Copy the entire spectrum at once  into the output array. */
       memcpy(local_part_spectra + (p - start_idx) * nlam,
-             this_part_spectra.data(), nlam * sizeof(double));
+             this_part_spectra.data(), nlam * sizeof(FLOAT));
 
       /* Reset the local spectra for this particle. */
       std::fill(this_part_spectra.begin(), this_part_spectra.end(), 0.0);
@@ -579,8 +579,8 @@ static void shifted_spectra_loop_ngp_omp(GridProps *grid_props,
  * @param nthreads: The number of threads to use.
  */
 void shifted_spectra_loop_ngp(GridProps *grid_props, Particles *parts,
-                              double *part_spectra, const int nthreads,
-                              const double c) {
+                              FLOAT *part_spectra, const int nthreads,
+                              const FLOAT c) {
 
   /* First get the grid indices for all particles. */
   get_particle_indices(grid_props, parts, nthreads);
@@ -664,9 +664,9 @@ PyObject *compute_part_seds_with_vel_shift(PyObject *self, PyObject *args) {
   RETURN_IF_PYERR();
 
   /* Allocate the spectra. */
-  double *spectra = new (std::nothrow) double[grid_props->nlam]();
-  double *part_spectra =
-      new (std::nothrow) double[npart * grid_props->nlam]();
+  FLOAT *spectra = new (std::nothrow) FLOAT[grid_props->nlam]();
+  FLOAT *part_spectra =
+      new (std::nothrow) FLOAT[npart * grid_props->nlam]();
 
   if (spectra == NULL || part_spectra == NULL) {
     PyErr_SetString(PyExc_MemoryError, "Failed to allocate memory for spectra.");
@@ -680,7 +680,7 @@ PyObject *compute_part_seds_with_vel_shift(PyObject *self, PyObject *args) {
   }
 
   /* Convert c to double */
-  double c = PyFloat_AsDouble(py_c);
+  FLOAT c = (FLOAT)PyFloat_AsDouble(py_c);
 
   toc("Extracting Python data", setup_start);
 
@@ -712,12 +712,12 @@ PyObject *compute_part_seds_with_vel_shift(PyObject *self, PyObject *args) {
   /* Construct the particle spectra output numpy array. */
   npy_intp np_dims[2] = {npart, nlam};
   PyArrayObject *out_part_spectra =
-      wrap_array_to_numpy<double>(2, np_dims, part_spectra);
+      wrap_array_to_numpy<FLOAT>(2, np_dims, part_spectra);
 
   /* Construct the integrated spectra output numpy array. */
   npy_intp np_dims_int[1] = {nlam};
   PyArrayObject *out_integrated_spectra =
-      wrap_array_to_numpy<double>(1, np_dims_int, spectra);
+      wrap_array_to_numpy<FLOAT>(1, np_dims_int, spectra);
 
   /* Construct the output tuple. */
   PyObject *out_tuple =
