+++ src/synthesizer/extensions/octree.h
@@ -8,6 +8,9 @@
 /* C headers. */
 #include <stdint.h>
 
+/* Local headers. */
+#include "data_types.h"
+
 /* Define the maximum tree depth. */
 #define MAX_DEPTH 64
 
@@ -17,13 +20,13 @@
 struct particle {
 
   /* Position of the particle. */
-  double pos[3];
+  FLOAT pos[3];
 
   /* Smoothing length of the particle. */
-  double sml;
+  FLOAT sml;
 
   /* Surface density variable. */
-  double surf_den_var;
+  FLOAT surf_den_var;
 
   /*! The index of the particle in the original array. */
   int index;
@@ -35,8 +38,8 @@ struct particle {
 struct cell {
 
   /* Location and width */
-  double loc[3];
-  double width;
+  FLOAT loc[3];
+  FLOAT width;
 
   /* Is it split? */
   int split;
@@ -49,7 +52,7 @@ struct cell {
   struct particle *particles;
 
   /* Store the square of the maximum smoothing length. */
-  double max_sml_squ;
+  FLOAT max_sml_squ;
 
   /* Pointers to cells below this one. */
   struct cell *progeny;
@@ -59,11 +62,11 @@ struct cell {
 };
 
 /* Prototypes. */
-void construct_cell_tree(const double *pos, const double *sml,
-                         const double *surf_den_val, const int npart,
+void construct_cell_tree(const FLOAT *pos, const FLOAT *sml,
+                         const FLOAT *surf_den_val, const int npart,
                          struct cell *root, int ncells, int maxdepth,
                          int min_count);
 void cleanup_cell_tree(struct cell *c);
-double min_projected_dist2(struct cell *c, double x, double y);
+FLOAT min_projected_dist2(struct cell *c, FLOAT x, FLOAT y);
 
 #endif // OCTREE_H_
