+++ src/synthesizer/extensions/reductions.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "data_types.h"
+
 // Reduction prototypes
-void reduce_spectra(double *spectra, double *part_spectra, int nlam, int npart,
+void reduce_spectra(FLOAT *spectra, FLOAT *part_spectra, int nlam, int npart,
                     int nthreads);
